backend/requirements.txt
fastapi[all]
SQLAlchemy
SQLAlchemy-Utils
alembic
uvicorn 
pytz
pydantic
inflect
jinja2
psycopg2-binary
greenlet
protobuf
python-decouple
gunicorn
python-multipart
pydantic
pydantic-settings
